requirements.txt
amqp==2.6.1
anyjson==0.3.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.13
django-allauth==0.39.1
django-celery-results==1.2.1
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.10
importlib-metadata==2.0.0
kombu==4.6.11
oauthlib==3.1.0
python3-openid==3.2.0
pytz==2020.1
redis==3.2.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.3.0
